refactor: moved useDOMReadyById to browser.ts
